+++ src/workspaces/macos-diag/MacosDiagDefenderTab.tsx
