+++ src/workspaces/macos-diag/MacosDiagEnvironmentBanner.tsx
