+++ src/workspaces/macos-diag/MacosDiagPackagesTab.tsx
