+++ src/workspaces/macos-diag/MacosDiagProfilesTab.tsx
