+++ src/workspaces/macos-diag/MacosDiagTabStrip.tsx
