+++ src/workspaces/macos-diag/MacosDiagUnifiedLogTab.tsx
