+++ src/workspaces/macos-diag/MacosDiagWorkspace.tsx
