+++ src/workspaces/macos-diag/macos-diag-store.ts
