+++ src/workspaces/macos-diag/types.ts
